refactor: extract browser action parsing
